Restructure files in lvgl-module
