+++ Modules/lvgl-module/include/lvgl/icons/launcher.h
