+++ Modules/lvgl-module/include/lvgl/icons/statusbar.h
