+++ Modules/lvgl-module/source/devices/keyboard.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <lvgl/lvgl_keyboard.h>
+#include <../include/lvgl/devices/keyboard.h>
 
 #include <tactility/drivers/keyboard.h>
