+++ Tactility/Source/app/grovesettings/GroveSettings.cpp
@@ -2,9 +2,9 @@
 
 #include <lvgl.h>
 
+#include <lvgl/icons/shared.h>
 #include <tactility/device.h>
 #include <tactility/drivers/grove.h>
-#include <lvgl/lvgl_icon_shared.h>
 
 #include <Tactility/Tactility.h>
 #include <Tactility/lvgl/Toolbar.h>
