+++ Tactility/Source/app/timedatesettings/TimeDateSettings.cpp
@@ -12,8 +12,8 @@
 
 #include <lvgl.h>
 
+#include <lvgl/icons/shared.h>
 #include <tactility/log.h>
-#include <lvgl/lvgl_icon_shared.h>
 
 namespace tt::app::timedatesettings {
