+++ Tactility/Source/app/timezone/TimeZone.cpp
@@ -1,5 +1,5 @@
-#include <lvgl/lvgl_fonts.h>
-#include <lvgl/lvgl_icon_shared.h>
+#include <lvgl/icons/shared.h>
+#include <lvgl/fonts.h>
 
 #include <Tactility/app/AppContext.h>
 #include <Tactility/app/AppManifest.h>
