+++ Tactility/Source/service/webserver/WebServerService.cpp
@@ -30,7 +30,7 @@
 #include <lv_screenshot.h>
 #endif
 
-#include <lvgl/lvgl_icon_statusbar.h>
+#include <lvgl/icons/statusbar.h>
 
 #include <atomic>
 #include <cctype>
